requirements_old.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Bottleneck==1.3.5
branca==0.6.0
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
chinese-calendar==1.8.0
click==8.0.4
cmdstanpy==1.1.0
colorama==0.4.6
colorlog==5.0.1
comm==0.1.2
contourpy==1.0.5
convertdate==2.3.2
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
easydev==0.12.0
entrypoints==0.4
ephem==4.1.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fitter==1.4.1
folium==0.14.0
fonttools==4.25.0
fqdn==1.5.1
hijri-converter==2.2.2
holidays==0.26
idna==3.4
importlib-resources==5.12.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.0
jupyter-events==0.5.0
jupyter_server==2.0.4
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.25.0
openpyxl==3.0.10
packaging==23.0
palettable==3.3.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
# pip==23.2.1
platformdirs==2.6.0
pooch==1.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
prophet==1.1.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyMeeus==0.5.11
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.1
Send2Trash==1.8.0
setuptools==68.0.0
setuptools-git==1.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
uri-template==1.2.0
urllib3==1.26.16
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wheel==0.38.4
widgetsnbextension==4.0.4
win-inet-pton==1.1.0
wincertstore==0.2
